+++ ohkami_lib/Cargo.toml
@@ -1,19 +1,20 @@
 [package]
-name          = "ohkami_lib"
-version       = "0.1.1"
-edition       = "2021"
-authors       = ["kanarus <kanarus786@gmail.com>"]
-description   = "Internal library for ohkami - intuitive and declarative web framework"
-documentation = "https://docs.rs/ohkami"
-homepage      = "https://crates.io/crates/ohkami"
-repository    = "https://github.com/kana-rus/ohkami"
-readme        = "../README.md"
-keywords      = ["async", "http", "web", "server", "framework"]
-categories    = ["asynchronous", "web-programming::http-server"]
-license       = "MIT"
+name             = "ohkami_lib"
+version          = "0.1.1"
+edition          = "2021"
+authors          = ["kanarus <kanarus786@gmail.com>"]
+description      = "Internal library for ohkami - intuitive and declarative web framework"
+documentation    = "https://docs.rs/ohkami"
+homepage         = "https://crates.io/crates/ohkami"
+repository       = "https://github.com/kana-rus/ohkami"
+readme           = "../README.md"
+keywords         = ["async", "http", "web", "server", "framework"]
+categories       = ["asynchronous", "web-programming::http-server"]
+license          = "MIT"
 
 [dependencies]
-serde         = { version = "1.0" }
+serde            = { version = "1.0" }
+percent-encoding = { version = "2.3" }
 
 [dev-dependencies]
-serde_derive  = { version = "1.0" }
+serde_derive     = { version = "1.0" }
