+++ ohkami_lib/src/lib.rs
@@ -10,8 +10,8 @@ mod slice;
 pub use slice::{Slice, CowSlice};
 
 mod percent_encoding;
-pub use percent_encoding::{percent_decode, percent_decode_utf8};
+pub use percent_encoding::{percent_encode, percent_decode, percent_decode_utf8};
 
 pub mod serde_utf8;
-#[cfg(test)] mod serde_utf8_test;
 
+pub mod serde_urlencoded;
