+++ ohkami_lib/src/serde_urlencoded.rs
@@ -1,0 +1,30 @@
+mod ser;
+mod de;
+
+#[cfg(test)]
+mod _test;
+
+
+#[derive(Debug)]
+pub struct Error(String);
+const _: () = {
+    impl std::fmt::Display for Error {
+        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+            f.write_str(&self.0)
+        }
+    }
+    impl std::error::Error for Error {}
+
+    impl serde::ser::Error for Error {
+        fn custom<T>(msg:T) -> Self where T:std::fmt::Display {
+            Self(msg.to_string())
+        }
+    }
+    impl serde::de::Error for Error {
+        fn custom<T>(msg:T) -> Self where T:std::fmt::Display {
+            Self(msg.to_string())
+        }
+    }
+};
+
+pub(crate) enum Infallible {}
